The default location of the node DB api URL has been updated to the latest as specified by Nathan.
The .ionconfig file is now created with the values found in the node DB,
provided they are present. If some values are not specified (ie. are
null) then the defaults in the ION config tool JSON files are used
instead. If all ionconfig parameters (sdr_config_flags, sdr_wm_size,
heap_words and wm_size) are set to 'null' in the database for a node, then that
node will not get an ionconfig file specified for it in the ION model
and will, hence, not have a the .ionconfig file generated unless the
user explicitly adds one. This commit addresses issue #67